Add config.sh to .gitignore
config.sh is a file that holds all of the environment variables for testing the bot. This should not be exposed to the public as each developer's config.sh holds their discord bot token and other protected tokens. This commit removes this file from the directory all-together to ensure the safety of their discord bot tokens.