Update ban appeal DM
The ban appeal process needs updated now and in the future. The dyno link no longer works so the user will need to contact one of the staff to start the appeal process.

Once the site is able to begin the appeal, process, this will need to be updated again.