Strip/Restore predefined roles on Exile/Unexile (#221)
Ticket: MOD-134
This PR adds the option to strip/restore predefined roles upon
exile/unexile of a user. Also changes type of `user` in the `unexile`
service from `discord.User` to `discord.Member`.
The database portion of this PR will also serve as the base for sticky
roles when joining/leaving the server in the future.