Make get_log_channel async (#353)
get_log_channel was called with `await` but not defined as `async`. 
This led to the exception `TypeError: object TextChannel can't be used
in 'await' expression` when called.
Defined `get_log_channel` as async. This change is consistent with all
other classes in the module which do use async.

Reverting `await` removal from `get_log_channel` in #312